requirements.txt
# medaka requirements. 
# Add comments to keep track of why we are using particular versions
biopython
h5py
keras
tensorflow
numpy
pysam

